adding in sweetalerts2 up to wishlist page, need to move towards items page

+++ server/public/scripts/controllers/user.controller.js
@@ -1,5 +1,9 @@
 myApp.controller('UserController', ['UserService', 'ListService', '$http', '$location', function (UserService, ListService, $http, $location) {
   var self = this;
+  self.newList = {
+    name: '',
+    description: ''
+  };
   console.log('UserController created');
 
   // handles log out
@@ -14,7 +18,7 @@ myApp.controller('UserController', ['UserService', 'ListService', '$http', '$loc
         self.userObject.userName = response.data.username;
         ListService.getLists(self.userObject.userId);
       } else {
-        console.log('Failure getting user');
+        console.log('Not Logged In');
         $location.path('/home');
       }
     });
@@ -31,21 +35,54 @@ myApp.controller('UserController', ['UserService', 'ListService', '$http', '$loc
 
   //list route handling
   self.startList = (newList) => {
-    self.newList.userId = self.userObject.userId;
-    //console.log('New wishlist as: ', self.newList);
-    self.startNewList = false;
-    ListService.startList(newList);
+    if (self.newList.name === '' || self.newList.description === '' ) {
+      swal({
+        title: 'Please name and describe your list!',
+        type: 'warning'
+      })
+    } else {
+      self.newList.userId = self.userObject.userId;
+      //console.log('New wishlist as: ', self.newList);
+      self.startNewList = false;
+      swal({
+        title: 'List Created!',
+        text: 'Lets add to it!',
+        type: 'info',
+        confirmButtonText: 'Thanks!'
+      });
+      ListService.startList(newList);
+    }
   }
 
   self.deleteList = (list) => {
     list.userId = self.userObject.userId;
     //console.log('Delete list: ', list);
-    ListService.deleteList(list);
+    swal({
+      title: 'Are you sure you want to delete this list??',
+      type: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!',
+      cancelButtonText: "Oops! No Way!"
+    }).then(function () {
+      swal(
+        'Deleted!',
+        'Your file has been deleted.',
+        'success'
+      )
+      ListService.deleteList(list);
+    });
   }
 
   self.saveList = (list) => {
     list.userId = self.userObject.userId;
     console.log('Editing list: ', list);
+    swal({
+      title: 'List Updated!',
+      type: 'info',
+      confirmButtonText: 'Thanks!'
+    });
     ListService.saveList(list);
   }
